small fixes
fixed parameter buffer datatype
fixed arguments handling (for target parameter option)
 reduced open/close device calls

+++ src/varikey/varikey_command.hpp
@@ -550,7 +550,7 @@ namespace varikey
             FUNCTION function;
             union
             {
-                uint buffer[100];
+                uint8_t buffer[100];
                 backlight_t backlight;
                 display_t display;
                 features_t features;
